+++ CHANGELOG.md
@@ -35,11 +35,16 @@
 - 💥 **BREAKING:** The deprecated `typed_settings.attrs.hooks` module
   has been removed.
 
+- ✨ Added a built-in `TSConverter` as an alternative for `cattrs`
+  (which is still supported and recommended).
+
+- ✨ `cattrs` is now an optional dependency.  You can install it with
+  `python -m pip install -U typed-settings[cattrs]`.
+
 - ✨ Added a dictionary loader.  This is useful for testing purposes.
 
 - 📦 Switched from [safety] to [pip-audit].
 
-[!16]: https://gitlab.com/sscherfke/typed-settings/-/merge_requests/16
 [#20]: https://gitlab.com/sscherfke/typed-settings/-/issues/20
 [#30]: https://gitlab.com/sscherfke/typed-settings/-/issues/30
 [#41]: https://gitlab.com/sscherfke/typed-settings/-/issues/41